devop: refactor dashboard
