+++ package.json
@@ -40,8 +40,8 @@
     "@coolwallets/wallet": "0.2.0-beta.1",
     "@ensdomains/dnsregistrar": "0.4.0",
     "@myetherwallet/mew-components": "0.5.32-beta",
-    "@sentry/browser": "6.1.0",
-    "@sentry/integrations": "6.1.0",
+    "@sentry/browser": "6.2.0",
+    "@sentry/integrations": "6.2.0",
     "@unstoppabledomains/resolution": "1.17.0",
     "@vue/babel-preset-app": "4.5.11",
     "@walletconnect/browser": "1.0.0",
@@ -56,7 +56,7 @@
     "apollo-utilities": "1.3.4",
     "bitbox02-api": "0.8.0",
     "clipboard-copy": "4.0.1",
-    "core-js": "3.8.3",
+    "core-js": "3.9.0",
     "graphql": "15.5.0",
     "graphql-anywhere": "4.2.7",
     "graphql-tag": "2.11.0",
@@ -91,7 +91,7 @@
     "@vue/eslint-config-prettier": "6.0.0",
     "@vue/test-utils": "1.1.3",
     "@xkeshi/vue-qrcode": "git+https://github.com/MyEtherWallet/vue-qrcode.git",
-    "apollo": "2.32.1",
+    "apollo": "2.32.5",
     "axios": "0.21.1",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.1.0",
@@ -131,7 +131,7 @@
     "remark-cli": "9.0.0",
     "remark-preset-lint-recommended": "5.0.0",
     "rskjs-util": "1.0.3",
-    "sass": "1.32.7",
+    "sass": "1.32.8",
     "sass-loader": "10.1.1",
     "sinon": "9.2.4",
     "store": "2.0.12",
